Добавлено восстановление пароля

+++ src/Application/eShopApplication.Application.AppData/AccountRole/Service/AccountRoleService.cs
@@ -2,6 +2,8 @@
 using eShopApplication.Application.AppData.Adverts.Repository;
 using eShopApplication.Contracts.AccountRole;
 using eShopApplication.Contracts.Adverts;
+using eShopApplication.Domain.AccountRole;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -26,7 +28,6 @@ namespace eShopApplication.Application.AppData.AccountRole.Service
         {
             var accountRole = new Domain.AccountRole.AccountRole
             {
-
                 AccountRoleName = createAccountRoleDto.AccountRoleName,
                 AccountRoleDescription = createAccountRoleDto.AccountRoleDescription
             };
@@ -37,6 +38,20 @@ namespace eShopApplication.Application.AppData.AccountRole.Service
         /// <inheritdoc cref="IAccountRoleService.DeleteAccountRoleAsync(Guid, CancellationToken)"/>
         public async Task DeleteAccountRoleAsync(Guid id, CancellationToken cancellationToken)
         {
+            var accountRoles = await _accountRoleRepository.GetAllAccountRolesAsync(cancellationToken);
+
+            var role = accountRoles.Select(s => new ReadAccountRoleDto
+            {
+                AccountRoleId = s.AccountRoleId,
+                AccountRoleName = s.AccountRoleName,
+                AccountRoleDescription = s.AccountRoleDescription
+            }).Where(s => s.AccountRoleId.Equals(id));
+
+            if(role == null)
+            {
+                throw new Exception($"Роль с идентификатором {id} не найдена.");
+            }
+
             await _accountRoleRepository.DeleteAccountRoleAsync(id, cancellationToken);
         }
